+++ drydock/duster.py
@@ -143,12 +143,12 @@ class MetaContainer(Container):
         cmd.append("--name " + self.fqdn)
         cmd.append("-h " + self.fqdn)
         cmd.extend(self.get_portmaps())
-        ngx = "-v /var/lib/{}/nginx/sites-enabled:/etc/nginx/sites-enabled"
+        ngx = "-v /var/lib/{}/nginx/sites-enabled:/etc/nginx/sites-enabled:ro"
         cmd.append(ngx.format(self.fqdn))
-        ngx = "-v /var/lib/{}/nginx/ssl:/etc/nginx/ssl"
+        ngx = "-v /var/lib/{}/nginx/ssl:/etc/nginx/ssl:ro"
         cmd.append(ngx.format(self.fqdn))
         cmd.append("-v /etc/timezone:/etc/timezone:ro")
-        cmd.append("-v /var/lib/{0}/drydock:/drydock".format(self.fqdn))
+        cmd.append("-v /var/lib/{0}/drydock:/drydock:ro".format(self.fqdn))
         cmd.extend(self.get_volumemaps())
         cmd.append(self.base)
         cmd.append("bash -l -c")